+++ client/src/screens/Workspaces/CUModal.test.tsx
@@ -3,24 +3,24 @@ import { render, fireEvent } from "@testing-library/react";
 import "@testing-library/jest-dom";
 
 import t from "../../utils/translation";
-import NewWorkspaceModal from "./NewWorkspaceModal";
+import CUModal from "./CUModal";
 
 beforeEach(() => {
   jest.restoreAllMocks();
 });
 
 test("renders without crashing", async () => {
-  await act(() => render(<NewWorkspaceModal show={true} />));
+  await act(() => render(<CUModal show={true} />));
 });
 
 test("hidden if not shown", async () => {
-  const modal = await act(() => render(<NewWorkspaceModal show={false} />));
+  const modal = await act(() => render(<CUModal show={false} />));
 
   expect(modal.queryByText(t("Titel*"))).toBeNull();
 });
 
 test("shows the expected inputs", async () => {
-  const modal = await act(() => render(<NewWorkspaceModal show={true} />));
+  const modal = await act(() => render(<CUModal show={true} />));
 
   const element = modal.queryByLabelText(t("Titel*"));
   expect(element).toBeInTheDocument();
@@ -29,7 +29,7 @@ test("shows the expected inputs", async () => {
 
 test("shows alert on bad inputs", async () => {
   // render
-  const modal = await act(() => render(<NewWorkspaceModal show={true} />));
+  const modal = await act(() => render(<CUModal show={true} />));
 
   expect(modal.queryByRole("alert")).not.toBeInTheDocument();
 
@@ -55,17 +55,17 @@ test("makes API-call on submission", async () => {
   jest.spyOn(global, "fetch").mockImplementation(fetchMock);
 
   // render
-  const modal = await act(() => render(<NewWorkspaceModal show={true} />));
+  const modal = await act(() => render(<CUModal show={true} />));
 
   // fill and submit form
   fireEvent.change(modal.getByLabelText(t("Titel*")), {
     target: { value: "Hello" },
   });
-  fireEvent.blur(modal.getByLabelText(t("Titel*")));  // trigger validation
+  fireEvent.blur(modal.getByLabelText(t("Titel*"))); // trigger validation
   fireEvent.click(modal.getByText(t("Erstellen")));
   await new Promise(process.nextTick);
 
   // assert
   expect(modal.queryByRole("alert")).not.toBeInTheDocument();
-  expect(fetchMock).toBeCalledTimes(2);  // submission + workspace-reload
+  expect(fetchMock).toBeCalledTimes(2); // submission + workspace-reload
 });